tenderflow-guinea/apps/api/requirements.txt
# TenderFlow Guinea — API Dependencies

# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0
alembic==1.13.2

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# HTTP
httpx==0.27.2

# Redis
redis==5.1.0
hiredis==3.0.0

# MinIO / S3
minio==7.2.9

# Document Processing
PyMuPDF==1.24.10
python-docx==1.1.2
beautifulsoup4==4.12.3
lxml==5.3.0
reportlab==4.2.2

# Celery
celery==5.4.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Utilities
python-dotenv==1.0.1
